# site2/backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: COLOR03 像素画转换 API
  description: |
    COLOR03 像素画转换 API 提供了完整的图像到像素画转换服务，支持多种插值算法、颜色量化方法和抖动处理。
    
    ## 主要特性
    - 🎨 多种插值算法（最近邻、双线性）
    - 🌈 高级颜色量化（八叉树、K-means）
    - ✨ 专业抖动处理（Bayer矩阵、Floyd-Steinberg）
    - 🚀 性能优化和内存管理
    - 🛡️ 完善的错误处理和重试机制
    
  version: 1.0.0
  contact:
    name: ColorMagic Team
    url: https://github.com/colormagic/pixel-art-api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3001/api/color03
    description: 开发环境
  - url: https://api.colormagic.com/color03
    description: 生产环境

paths:
  /health:
    get:
      summary: 健康检查
      description: 检查服务健康状态和可用性
      operationId: healthCheck
      tags:
        - 系统监控
      responses:
        '200':
          description: 服务健康
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
              example:
                success: true
                status: "healthy"
                timestamp: "2024-01-20T10:30:00.000Z"
                checkId: "health_1641823800000"
                services:
                  sharp:
                    status: true
                    error: null
                  memory:
                    status: true
                    warning: false
                    usage:
                      heapUsed: "45.67MB"
                      heapTotal: "89.23MB"
                      external: "12.34MB"
                      rss: "156.78MB"
                  uptime:
                    status: true
                    minutes: "123.45"
        '503':
          description: 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /pixel-art/convert:
    post:
      summary: 像素画转换
      description: 将上传的图像转换为像素画
      operationId: convertToPixelArt
      tags:
        - 图像处理
      parameters:
        - name: quality
          in: query
          description: 处理质量级别
          required: false
          schema:
            type: string
            enum: [fast, balanced, high_quality]
            default: balanced
          example: balanced
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - imageFile
                - resizeFactor
                - interpolation
                - colorMode
                - ditheringRatio
              properties:
                imageFile:
                  type: string
                  format: binary
                  description: 图像文件 (PNG/JPG/JPEG, 最大10MB)
                resizeFactor:
                  type: number
                  minimum: 0
                  maximum: 100
                  description: 缩放因子 (0-100)
                  example: 25
                interpolation:
                  type: string
                  enum: [nearest_neighbor, bilinear]
                  description: 插值方法
                  example: nearest_neighbor
                colorMode:
                  type: string
                  enum: [no_dithering, ordered_dithering_bayer]
                  description: 颜色模式
                  example: ordered_dithering_bayer
                ditheringRatio:
                  type: number
                  minimum: 0.1
                  maximum: 3.0
                  description: 抖动比例 (0.1-3.0)
                  example: 1.5
            encoding:
              imageFile:
                contentType: image/png, image/jpeg, image/jpg
      responses:
        '200':
          description: 转换成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversionSuccessResponse'
              example:
                success: true
                data:
                  pixelArtImage: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                  canvasInfo:
                    width: 64
                    height: 48
                    coloredPixels: 3072
                  extractedColors:
                    - "#FF0000"
                    - "#00FF00"
                    - "#0000FF"
                    - "#FFFFFF"
                    - "#000000"
                  processingTime: 1250.67
                  requestId: "req_1641823800000_abc123def"
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  summary: 参数验证错误
                  value:
                    success: false
                    error: "输入参数有误，建议：检查输入参数范围"
                    code: "VALIDATION_ERROR"
                    suggestions:
                      - "检查输入参数范围"
                      - "验证必需参数"
                      - "使用默认参数重试"
                image_too_large:
                  summary: 图像过大
                  value:
                    success: false
                    error: "图像尺寸过大，最大支持 4096x4096 像素"
                    code: "IMAGE_TOO_LARGE"
                    suggestions:
                      - "使用更小的图像"
                      - "裁剪图像到支持的尺寸"
        '408':
          description: 处理超时
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: 文件过大或内存不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "内存不足，建议：启用分块处理模式"
                code: "MEMORY_LIMIT_EXCEEDED"
                suggestions:
                  - "启用分块处理模式"
                  - "减少图像尺寸"
                  - "降低颜色数量"
        '422':
          description: 处理失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    HealthCheckResponse:
      type: object
      required:
        - success
        - status
        - timestamp
        - checkId
      properties:
        success:
          type: boolean
          description: 请求是否成功
        status:
          type: string
          enum: [healthy, unhealthy, error]
          description: 服务状态
        timestamp:
          type: string
          format: date-time
          description: 检查时间戳
        checkId:
          type: string
          description: 检查ID
        services:
          type: object
          description: 各服务状态详情
          properties:
            sharp:
              type: object
              properties:
                status:
                  type: boolean
                error:
                  type: string
                  nullable: true
            memory:
              type: object
              properties:
                status:
                  type: boolean
                warning:
                  type: boolean
                usage:
                  type: object
                  properties:
                    heapUsed:
                      type: string
                    heapTotal:
                      type: string
                    external:
                      type: string
                    rss:
                      type: string
            uptime:
              type: object
              properties:
                status:
                  type: boolean
                minutes:
                  type: string
        detailed:
          type: object
          description: 详细系统信息
          properties:
            memory:
              type: object
              description: 详细内存使用信息
            uptime:
              type: string
            nodeVersion:
              type: string
            platform:
              type: string
            arch:
              type: string

    ConversionSuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - pixelArtImage
            - canvasInfo
            - extractedColors
            - processingTime
            - requestId
          properties:
            pixelArtImage:
              type: string
              description: Base64编码的PNG图像
              example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
            canvasInfo:
              type: object
              required:
                - width
                - height
                - coloredPixels
              properties:
                width:
                  type: integer
                  description: 输出图像宽度
                  example: 64
                height:
                  type: integer
                  description: 输出图像高度
                  example: 48
                coloredPixels:
                  type: integer
                  description: 有颜色的像素数量
                  example: 3072
            extractedColors:
              type: array
              items:
                type: string
                pattern: '^#[0-9A-Fa-f]{6}$'
              description: 提取的调色板颜色数组
              example: ["#FF0000", "#00FF00", "#0000FF"]
            processingTime:
              type: number
              description: 处理时间（毫秒）
              example: 1250.67
            requestId:
              type: string
              description: 请求追踪ID
              example: "req_1641823800000_abc123def"

    ErrorResponse:
      type: object
      required:
        - success
        - error
        - code
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: 用户友好的错误消息
          example: "图像尺寸过大，最大支持 4096x4096 像素"
        code:
          type: string
          description: 错误代码
          enum:
            - VALIDATION_ERROR
            - INVALID_IMAGE_CONTENT
            - IMAGE_TOO_LARGE
            - IMAGE_TOO_MANY_PIXELS
            - MEMORY_LIMIT_EXCEEDED
            - PROCESSING_TIMEOUT
            - CONVERSION_FAILED
            - INTERNAL_ERROR
          example: "IMAGE_TOO_LARGE"
        suggestions:
          type: array
          items:
            type: string
          description: 解决建议
          example:
            - "使用更小的图像"
            - "裁剪图像到支持的尺寸"
        requestId:
          type: string
          description: 请求追踪ID
          example: "req_1641823800000_abc123def"

  examples:
    BasicConversionRequest:
      summary: 基础转换请求
      description: 使用最近邻插值和无抖动的基础像素画转换
      value:
        resizeFactor: 25
        interpolation: "nearest_neighbor"
        colorMode: "no_dithering"
        ditheringRatio: 1.0

    HighQualityConversionRequest:
      summary: 高质量转换请求
      description: 使用双线性插值和Bayer矩阵抖动的高质量转换
      value:
        resizeFactor: 50
        interpolation: "bilinear"
        colorMode: "ordered_dithering_bayer"
        ditheringRatio: 2.0

tags:
  - name: 图像处理
    description: 像素画转换相关API
  - name: 系统监控
    description: 服务状态监控API

externalDocs:
  description: 完整API文档
  url: ./API_SPECIFICATION.md